constants cleaning and in proper file
