+++ src/renderBoard.ts
@@ -1,13 +1,6 @@
 import { Board, Player } from "./types";
 
-export const RED_MARBLE: string = `\u001b[31m \u2022 \u001b[0m`;
-export const BLUE_MARBLE: string = `\u001b[34m \u2022 \u001b[0m`;
-export const WHITE_MARBLE: string = `\u001b[37m \u2022 \u001b[0m`;
-export const EMPTY_MARBLE: string = `   `;
-const MARBLE_COLORS = [EMPTY_MARBLE, RED_MARBLE, BLUE_MARBLE, WHITE_MARBLE];
-
-export const ALPHABET: string = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
-const MARBLE_INT_COLORS: string[] = ["empty", "red", "blue", "white"];
+import { MARBLE_COLORS, ALPHABET, MARBLE_INT_COLORS } from "./constants";
 
 export function renderBoard(board: Board): string {
   const columnLetters = ALPHABET.substr(0, board[0].length);
